refactor(frontend): decouple productivity logic into viewmodel and extract EventCard component
Decomposes the productivity "God Class" by extracting business logic, complex state, and memoized list calculations into a dedicated `useProductivityViewModel` hook. Extracts inline `EventCard` view logic into a separate component. Pulls `renderItem` outside of the main component body for FlatList optimization. Adheres to clean architecture and SOLID principles.
